small restructure + added logging
