+++ tempusEdge-opcMonitor/src/main/scala/com/hashmapinc/opcMonitor/Config.scala
@@ -1,0 +1,19 @@
+package com.hashmapinc.opcMonitor
+
+object Config {
+  // general configs
+  val context = "production"
+
+  // MQTT configs
+
+  // OPC configs 
+
+  /**
+   * This function sets the configs to test values
+   */
+  def initTestContext {
+    // MQTT configs
+
+    // OPC configs
+  }
+}
