+++ tempusEdge-opcMonitor/src/main/scala/com/hashmapinc/opcMonitor/iofog/IofogController.scala
@@ -1,0 +1,5 @@
+package com.hashmapinc.opcMonitor.iofog
+
+object IofogController {
+
+}
