+++ tempusEdge-opcMonitor/src/main/scala/com/hashmapinc/opcMonitor/mqtt/MqttController.scala
@@ -1,0 +1,12 @@
+package com.hashmapinc.opcMonitor.mqtt
+
+//import paho
+import org.eclipse.paho.client.mqttv3.MqttClient;
+import org.eclipse.paho.client.mqttv3.MqttConnectOptions;
+import org.eclipse.paho.client.mqttv3.MqttException;
+import org.eclipse.paho.client.mqttv3.MqttMessage;
+import org.eclipse.paho.client.mqttv3.persist.MemoryPersistence;
+
+object MqttController {
+
+}
